[Editorial] Fix inconsistent use of "onto" vs "on to"
The 1917 edition of the book removes the first "on" in the sentence
"You have to go ’cross the river an’ get on to our side if you want to
see anything goin’ on."

This sidesteps the issue of having to decide which spelling to use
"onto" or "on to" to maintain consistent spelling.

Previously in chapter 1, a character said: “Now you watch me heave this
newspaper right onto Mis’ Brown’s doorstep.”

There are a handful of other editions that use this word choice
(1931, 1970,1981, 1984, 1988).